added cpuinfo and poweroff but it is not working currently

+++ kernel/src/interrupts/timer.h
@@ -8,7 +8,7 @@ void init_timer(void);
 void sleep(uint32_t ticks);
 void sleep_s(uint32_t seconds);
 void sleep_m(uint32_t minutes);
-int get_tick_count();
+uint64_t get_tick_count();
 
 
 #endif // TIMER_H
